Intégration de MULTER.

+++ controllers/library.js
@@ -1,10 +1,17 @@
+const { error } = require("console");
 const Book = require("../models/Book");
+const fs = require("fs");
 
 exports.createBook = (req, res, next) => {
-	delete req.body._id;
+	const bookObject = JSON.parse(req.body.book);
+	delete bookObject._id;
+	delete bookObject._userId;
 	const book = new Book({
-		...req.body
+		...bookObject,
+		userId: req.auth.userId,
+		imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`
 	});
+
 	book.save()
 		.then(() => res.status(201).json({ message: "Objet enregistré !" }))
 		.catch(error => res.status(400).json({ error }));
@@ -23,13 +30,46 @@ exports.getAllBooks = (req, res, next) => {
 };
 
 exports.modifyBook = (req, res, next) => {
-	Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-		.then(() => res.status(200).json({ message: "Objet modifié !" }))
+	const bookObject = req.file ? {
+		...JSON.parse(req.body.book),
+		imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`
+	} : {...req.body};
+
+	delete bookObject._userId;
+	Book.findOne({_id: req.params.id})
+		.then((book) => {
+			if (book.userId != req.auth.userId) {
+				res.status(401).json({ message: "Non-autorisé" });
+			} else {
+				// Suppresion de l'ancienne image si une nouvelle est insérée dans la modification
+				if (req.file) {
+					const oldFilename = book.imageUrl.split("/images/")[1];
+					fs.unlink(`images/${oldFilename}`, (error) => {
+						if (error) res.status(500).json({ error });
+					});
+				};
+				// 
+				Book.updateOne({ _id: req.params.id }, { ...bookObject, _id: req.params.id })
+					.then(() => res.status(200).json({ message: "Objet modifié !" }))
+					.catch(error => res.status(401).json({ error }));
+			};
+		})
 		.catch(error => res.status(400).json({ error }));
 };
 
 exports.deleteBook = (req, res, next) => {
-	Book.deleteOne({ _id: req.params.id })
-		.then(() => res.status(200).json({ message: "Objet supprimé !" }))
-		.catch(error => res.status(400).json({ error }));
+	Book.findOne({ _id: req.params.id })
+		.then(book => {
+			if (book.userId != req.auth.userId) {
+				res.status(401).json({ message: "Non-autorisé" })
+			} else {
+				const filename = book.imageUrl.split("/images/")[1];
+				fs.unlink(`images/${filename}`, () => {
+					Book.deleteOne({ _id: req.params.id })
+						.then(() => res.status(200).json({ message: "Objet supprimé !" }))
+						.catch(error => res.status(401).json({ error }));
+				});
+			};
+		})
+		.catch(error => res.status(500).json({ error }));
 };
